migrations update: numeric to integer

+++ db/migrate/20171116150305_create_player_teams.rb
@@ -1,8 +1,8 @@
 class CreatePlayerTeams < ActiveRecord::Migration[5.1]
   def change
     create_table :player_teams do |t|
-      t.numeric :goals
-      t.numeric :faults
+      t.integer :goals
+      t.integer :faults
 
       t.timestamps
     end
